requirements.txt
numpy
openai==0.27.8
pynput==1.7.6
python-dotenv==1.0.0
scipy==1.8.0
sounddevice==0.4.6